refactor: 유즈케이스 프로토콜 각각의 파일로 분리 및 관련성에 따라 그룹화
