+++ iOS/FlipMate/FlipMate/Domain/UseCases/Protocol/Authentication/SignOutUseCase.swift
@@ -1,0 +1,12 @@
+//
+//  SignOutUseCase.swift
+//  FlipMate
+//
+//  Created by 권승용 on 1/7/24.
+//
+
+import Foundation
+
+protocol SignOutUseCase {
+    func signOut()
+}
