+++ iOS/FlipMate/FlipMate/Domain/UseCases/Protocol/Category/DeleteCategoryUseCase.swift
@@ -1,0 +1,12 @@
+//
+//  DeleteCategoryUseCase.swift
+//  FlipMate
+//
+//  Created by 권승용 on 1/7/24.
+//
+
+import Foundation
+
+protocol DeleteCategoryUseCase {
+    func deleteCategory(of id: Int) async throws
+}
