+++ iOS/FlipMate/FlipMate/Domain/UseCases/Protocol/Category/ReadCategoryUseCase.swift
@@ -1,0 +1,12 @@
+//
+//  ReadCategoryUseCase.swift
+//  FlipMate
+//
+//  Created by 권승용 on 1/7/24.
+//
+
+import Foundation
+
+protocol ReadCategoryUseCase {
+    func readCategory() async throws -> [Category]
+}
