+++ iOS/FlipMate/FlipMate/Domain/UseCases/Protocol/Category/UpdateCategoryUseCase.swift
@@ -1,0 +1,12 @@
+//
+//  UpdateCategoryUseCase.swift
+//  FlipMate
+//
+//  Created by 권승용 on 1/7/24.
+//
+
+import Foundation
+
+protocol UpdateCategoryUseCsae {
+    func updateCategory(of id: Int, newName: String, newColorCode: String) async throws
+}
