+++ iOS/FlipMate/FlipMate/Domain/UseCases/Protocol/Chart/FetchDailyChartUseCase.swift
@@ -10,7 +10,3 @@ import Foundation
 protocol FetchDailyChartUseCase {
     func fetchDailyChartLog(at date: Date) async throws -> CategoryChartLog
 }
-
-protocol FetchWeeklyChartUseCase {
-    func fetchWeeklyChartLog() async throws -> WeeklyChartLog
-}
