+++ iOS/FlipMate/FlipMate/Domain/UseCases/Protocol/Chart/FetchWeeklyChartUseCase.swift
@@ -1,0 +1,12 @@
+//
+//  FetchWeeklyChartUseCase.swift
+//  FlipMate
+//
+//  Created by 권승용 on 1/7/24.
+//
+
+import Foundation
+
+protocol FetchWeeklyChartUseCase {
+    func fetchWeeklyChartLog() async throws -> WeeklyChartLog
+}
