+++ iOS/FlipMate/FlipMate/Domain/UseCases/Protocol/Friend/SearchFriendUseCase.swift
@@ -1,0 +1,13 @@
+//
+//  SearchFriendUseCase.swift
+//  FlipMate
+//
+//  Created by 권승용 on 1/7/24.
+//
+
+import Foundation
+import Combine
+
+protocol SearchFriendUseCase {
+    func search(at nickname: String) -> AnyPublisher<FriendSearchResult, NetworkError>
+}
