+++ iOS/FlipMate/FlipMate/Domain/UseCases/Protocol/Friend/UnfollowFriendUseCase.swift
@@ -1,0 +1,13 @@
+//
+//  UnfollowFriendUseCase.swift
+//  FlipMate
+//
+//  Created by 권승용 on 1/7/24.
+//
+
+import Foundation
+import Combine
+
+protocol UnfollowFriendUseCase {
+    func unfollow(at id: Int) -> AnyPublisher<String, NetworkError>
+}
