+++ iOS/FlipMate/FlipMate/Domain/UseCases/Protocol/GetStudyLogUseCase.swift
@@ -1,0 +1,13 @@
+//
+//  StudyLogUseCase.swift
+//  FlipMate
+//
+//  Created by 임현규 on 2023/11/23.
+//
+
+import Foundation
+import Combine
+
+protocol GetStudyLogUseCase {
+    func getStudyLog() -> AnyPublisher<StudyLog, NetworkError>
+}
