+++ iOS/FlipMate/FlipMate/Domain/UseCases/Protocol/Social/GetFriendsUseCase.swift
@@ -1,0 +1,13 @@
+//
+//  SocialUseCase.swift
+//  FlipMate
+//
+//  Created by 임현규 on 2023/11/30.
+//
+
+import Foundation
+import Combine
+
+protocol GetFriendsUseCase {
+    func getMyFriend(date: Date) -> AnyPublisher<[Friend], NetworkError>
+}
