+++ iOS/FlipMate/FlipMate/Domain/UseCases/Protocol/Social/SetupProfileInfoUseCase.swift
@@ -7,10 +7,6 @@
 
 import Foundation
 
-protocol ValidateNicknameUseCase {
-    func isNickNameValid(_ nickName: String) async throws -> NickNameValidationState
-}
-
 protocol SetupProfileInfoUseCase {
     func setupProfileInfo(nickName: String, profileImageData: Data) async throws -> UserInfo
 }
