+++ iOS/FlipMate/FlipMate/Domain/UseCases/Protocol/Timer/FinishTimerUseCase.swift
@@ -1,0 +1,13 @@
+//
+//  FinishTimerUseCase.swift
+//  FlipMate
+//
+//  Created by 권승용 on 1/7/24.
+//
+
+import Foundation
+import Combine
+
+protocol FinishTimerUseCase {
+    func finishTimer(endTime: Date, learningTime: Int, categoryId: Int?) -> AnyPublisher<Void, NetworkError>
+}
